requirements.txt
beautifulsoup4==4.5.1
click==6.7
cookies==2.2.1
Flask==0.11.1
Flask-Cors==2.1.2
geopy==1.11.0
googlefinance==0.7
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
oauth2==1.9.0.post1
oauthlib==2.0.1
python-dateutil==2.5.3
python-dotenv==0.6.2
python-forecastio==1.3.5
python-instagram==1.3.2
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
Werkzeug==0.11.11
yelp==1.0.2